filters working. need to add implementation for redirecting to Product Compare/Details

+++ MarketplaceWebPortal.Core/POCOClasses/ProductDetailUI.cs
@@ -39,8 +39,9 @@ namespace MarketplaceWebPortal.Core.POCOClasses
         
         //technical specifications
         public int air_flow { get; set; }
-        public decimal power_min { get; set; }
-        public decimal power_max { get; set; }
+        //changed to double because of .net package reference error
+        public double power_min { get; set; }
+        public double power_max { get; set; }
         public int operating_voltage_min { get; set; }
         public int operating_voltage_max { get; set; }
         public int fan_speed_min { get; set; }
@@ -49,9 +50,9 @@ namespace MarketplaceWebPortal.Core.POCOClasses
         public int sound_at_max_speed { get; set; }
 
         public float fan_sweep_diameter { get; set; }
-        public decimal height_min { get; set; }
-        public decimal height_max { get; set; }
-        public decimal weight { get; set; }
+        public double height_min { get; set; }
+        public double height_max { get; set; }
+        public double weight { get; set; }
 
         
         public string Accessories { get; set; }
